+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bugfix
 
 - Fix blurhash canvas scaling when image has a css aspect-ratio set #276 @reebalazs
+- Always clean extra properties, even if not generating a srcset #277 @reebalazs
 
 ### Internal